requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-appconf==1.0.4
django-autocomplete-light==3.5.1
django-bootstrap-form==3.4
django-celery-results==1.2.1
django-compressor==2.4
django-filter==2.2.0
django-polymorphic==2.1.2
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
html5lib==1.0.1
icalendar==4.0.6
idna==2.9
jsonify==0.5
kombu==4.6.8
Pillow==7.1.2
psycopg2==2.8.5
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
rcssmin==1.0.6
redis==3.5.1
reportlab==3.5.42
requests==2.23.0
rjsmin==1.1.0
schedule==0.6.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
